Get rid of HrmFormController
